+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Go server: Add support for `labeled_boolean` metrics with static labels ([AE-1250](https://mozilla-hub.atlassian.net/browse/AE-1250))
 - BUGFIX: Go server: Serialize nil `string_list` metrics as empty arrays instead of null ([#837](https://github.com/mozilla/glean_parser/pull/837))
+- Add support to target `glean_sym` (the new experimental Glean Rust API) ([#841](https://github.com/mozilla/glean_parser/pull/841))
 
 ## 19.0.0